[Silver V] Title: 귀찮음, Time: 180 ms, Memory: 8176 KB, Score: 30 point -BaekjoonHub

+++ 백준/Silver/16208. 귀찮음/귀찮음.cc
@@ -6,17 +6,18 @@ using namespace std;
 int main() {
     int n;
     cin >> n;
-    priority_queue<int, vector<int>, greater<int>> pq;
-    int result = 0;
+    priority_queue<long long, vector<long long>, greater<long long>> pq;
+    long long result = 0;
 
     for (int i=0; i<n; i++) {
         int tmp;
         cin >> tmp;
         pq.push(tmp);
     }
+    //(50,500,000 - 101) * 101
 
     for (int i=0; i<n-1; i++) {
-        int a, b;
+        long long a, b;
         a = pq.top();
         pq.pop();
         b = pq.top();
